refactor(design-system): remove deprecated components
