+++ libs/design-system/package.json
@@ -44,7 +44,6 @@
     "yup": "^0.32.11",
     "react-qr-code": "^2.0.7",
     "next-themes": "^0.2.1",
-    "date-fns": "^2.28.0",
     "d3-array": "^3.1.6",
     "clipboard-polyfill": "^4.0.1",
     "react-dropzone": "^14.2.3",
