+++ libs/design-system/src/index.ts
@@ -79,6 +79,7 @@ export * from './components/EmptyState/StateNoneYet'
 export * from './components/EmptyState/StateNoData'
 export * from './components/EmptyState/StateNoneMatching'
 export * from './components/EmptyState/StateError'
+export * from './components/WebDomain'
 
 // app
 export * from './app/AppPublicLayout'
@@ -112,6 +113,8 @@ export * from './app/PanelMenuSetting'
 export * from './app/ConfirmDialog'
 export * from './app/DaemonProfile'
 export * from './app/TestnetWarningBanner'
+export * from './app/NextAppSsrAppRouter'
+export * from './app/NextAppCsr'
 
 // form
 export * from './form/ConfigurationPanel'
@@ -138,21 +141,6 @@ export * from './form/useFormChangeCount'
 export * from './form/useFormInit'
 export * from './form/useFormSetField'
 export * from './form/useDialogFormHelpers'
-
-// site
-export * from './site/SiteHeading'
-export * from './site/SiteMap'
-export * from './site/ContentGallery'
-export * from './site/ContentItem'
-export * from './site/ContentProject'
-export * from './site/TransparentGradient'
-export * from './site/PatternedPanel'
-export * from './site/Callout'
-export * from './site/Links'
-export * from './site/NextAppSsrAppRouter'
-export * from './site/NextAppCsr'
-export * from './site/BackgroundImagePattern'
-export * from './site/WebDomain'
 
 // hooks
 export * from './hooks/tooltip'
